# specs/api.yaml
openapi: 3.1.0
info:
  title: PsiZero Resonance Platform API
  version: 1.0.0
  description: >
    Integrated APIs for symbolic resonance technologies:
    - SRS (Symbolic Resonance Solver) for NP-complete problems
    - HQE (Holographic Quantum Encoder) simulator
    - QSEM (Quantum Semantics) for concept resonance
    - NLC (Non-Local Communication) prime-eigenstate channels
    - QCR (Quantum Consciousness Resonator) sessions
    - I-CHING oracle (hexagram evolution)
    - Unified resonance physics (observer-capacity/gravity model)
servers:
  - url: https://api.psizero.dev
    description: Production
  - url: https://sandbox.psizero.dev
    description: Sandbox
...
security:
  - ApiKeyAuth: []
  - OAuth2CC: [platform.read, platform.write]
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    OAuth2CC:
      type: oauth2
      description: Client Credentials flow (for enterprises)
      flows:
        clientCredentials:
          tokenUrl: https://auth.psizero.dev/oauth/token
          scopes:
            platform.read: Read access
            platform.write: Write/access compute
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Provide to make POST requests idempotent (UUID recommended).
      schema: { type: string, maxLength: 128 }
...
  responses:
    ProblemJson:
      description: RFC 7807 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        meta: { type: object, additionalProperties: true }
...
    ResonanceMetrics:
      type: object
      properties:
        entropy: { type: number }
        plateauDetected: { type: boolean }
        dominance: { type: number, description: "max mass / coherence, [0,1]" }
        locality: { type: number, description: "non-locality inverse; lower=more non-local" }
        resonanceStrength: { type: number, description: "phase-lock strength, [0,1]" }

    # ---------- SRS ----------
...
    NLCSessionCreate:
      type: object
      properties:
        primes:
          type: array
          items: { type: integer }
        phases:
          type: array
          items: { type: number }
        goldenPhase: { type: boolean, default: true }
        silverPhase: { type: boolean, default: true }
      required: [primes]
    NLCSession:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [initializing, syncing, stable, degraded, closed] }
        metrics: { $ref: '#/components/schemas/ResonanceMetrics' }
    NLCMessage:
      type: object
      properties:
        content: { type: string }
        stamp: { type: string, format: date-time }
        channelQuality: { type: number }

    # ---------- QCR ----------
...
paths:
...
  /v1/nlc/sessions:
    post:
      tags: [NLC]
      summary: Create a non-local communication session
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NLCSessionCreate' }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NLCSession' }
        '400': { $ref: '#/components/responses/ProblemJson' }
  /v1/nlc/sessions/{id}:
    get:
      tags: [NLC]
      summary: Get NLC session status/metrics
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NLCSession' }
        '404': { $ref: '#/components/responses/ProblemJson' }
  /v1/nlc/sessions/{id}/messages:
    post:
      tags: [NLC]
      summary: Send a message over the resonance channel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, maxLength: 4000 }
      responses:
        '202':
          description: Accepted for transmission
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NLCMessage' }
        '404': { $ref: '#/components/responses/ProblemJson' }

  # ---------- QCR ----------
...